# Detected CU-Fallback case and then overwrite xcc values.