--- a/tensilelite/Tensile/Source/lib/include/Tensile/AMDGPU.hpp
+++ b/tensilelite/Tensile/Source/lib/include/Tensile/AMDGPU.hpp
@@ -2,7 +2,7 @@
  *
  * MIT License
  *
- * Copyright (C) 2022-2023 Advanced Micro Devices, Inc. All rights reserved.
+ * Copyright (C) 2022-2025 Advanced Micro Devices, Inc. All rights reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -231,8 +231,10 @@ namespace TensileLite
         int         skGridMultiplier = 1;
         int         skFixedGrid      = 0;
         int         skFullTiles      = 1;
+        mutable int isStandardCUs    = -1; // -1: unset, 0:false, 1:true
         std::string deviceName;
 
+        virtual bool   isStandardCU() const;
         virtual bool   runsKernelTargeting(Processor p) const;
         virtual size_t id() const
         {