--- a/tensilelite/Tensile/Source/lib/include/Tensile/Tensile.hpp
+++ b/tensilelite/Tensile/Source/lib/include/Tensile/Tensile.hpp
@@ -2,7 +2,7 @@
  *
  * MIT License
  *
- * Copyright (C) 2022 Advanced Micro Devices, Inc. All rights reserved.
+ * Copyright (C) 2022-2025 Advanced Micro Devices, Inc. All rights reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -163,8 +163,9 @@ namespace TensileLite
     public:
         virtual ~Solution();
 
-        virtual std::string name() const        = 0;
-        virtual std::string description() const = 0;
+        virtual std::string name() const                           = 0;
+        virtual std::string description() const                    = 0;
+        virtual bool        isFallbackForHW(Hardware const&) const = 0;
     };
 
     /**