Resolved:
- SWDEV-528780
- SWDEV-522233
- SWDEV-533436
------------
#2104 is our first workaround idea. But that PR is huge and hard to rebase in the meantime. And hard to maintain/restore in the near future.

This PR is another idea to dynamically overwrite xcc value when we know the solution is a CU-fallback. Should be a better solution and no need to modify any logic yamls.

Will re-run the tests in the three tickets.
- [x] SWDEV-528780 = SWDEV-522233 (swizzleA) - `./hipblaslt-test --gtest_filter=*matmul_swizzleA*`
- [x] SWDEV-533436 (nightly) - `./hipblaslt-test --gtest_filter=*heuristic_all_solutions*` and then `./hipblaslt-test --gtest_filter=*nightly*`
- [x] Additional test - `./hipblaslt-test --gtest_filter=*groupedgemm*`
- [x] General gtest

The whole testing can be faster on those non-standard CU devices, since lots of "false rejections" happened before this PR due to the XCC predicate. 

----------
### Test Results: (using non-standard CUs)
#### SWDEV-528780 = SWDEV-522233 : filter = matmul_swizzleA
[==========] 384 tests from 1 test suite ran. (27602 ms total)
[  PASSED  ] 384 tests.
hipBLASLt version: 1500
hipBLASLt git version: f52adfb7-dirty
command line: ./hipblaslt-test --gtest_filter=*matmul_swizzleA* 

#### filter = heuristic_all_solutions
[==========] 736 tests from 1 test suite ran. (1241772 ms total)
[  PASSED  ] 736 tests.
hipBLASLt version: 1500
hipBLASLt git version: 3a712cba-dirty
command line: ./hipblaslt-test --gtest_filter=*heuristic_all_solutions* 

#### SWDEV-533436: filter = nightly
[==========] 5272 tests from 1 test suite ran. (2586325 ms total)
[  PASSED  ] 5272 tests.
hipBLASLt version: 1500
hipBLASLt git version: 65338c70-dirty
command line: ./hipblaslt-test --gtest_filter=*nightly* 

#### General gtest
[==========] 54611 tests from 13 test suites ran. (12799729 ms total)
[  PASSED  ] 54611 tests.
hipBLASLt version: 1500
hipBLASLt git version: be6bdced-dirty
command line: ./hipblaslt-test

----------
### Additional Improvement - get solution will be faster on devices with specialized CUs:
#### Single grouped gemm test: 
**Before:**
[==========] 2 tests from 1 test suite ran. (**53014 ms total**)
[  PASSED  ] 2 tests.
hipBLASLt version: 1500
hipBLASLt git version: 3a712cba-dirty
command line: ./hipblaslt-test --gtest_filter=*matmul_groupedgemm_f8_fnuz* 

**After:**
[==========] 2 tests from 1 test suite ran. (**27690 ms total**)
[  PASSED  ] 2 tests.
hipBLASLt version: 1500
hipBLASLt git version: 65338c70-dirty
command line: ./hipblaslt-test --gtest_filter=*matmul_groupedgemm_f8_fnuz* 

This is because lots of "false rejections" happened before this PR due to the XCC predicate. This PR now can find more supported solutions with the ability of cu-fallback detection.